+++ previous-versions/assess/src/codatassess/models/operations/get_commerce_lifetime_value_metrics.py
@@ -36,7 +36,7 @@ class GetCommerceLifetimeValueMetricsResponse:
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     report: Optional[shared_report.Report] = dataclasses.field(default=None)