+++ previous-versions/assess/src/codatassess/models/operations/list_data_type_data_integrity_details.py
@@ -36,7 +36,7 @@ class ListDataTypeDataIntegrityDetailsResponse:
     details: Optional[shared_details.Details] = dataclasses.field(default=None)
     r"""OK"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""