+++ previous-versions/assess/src/codatassess/models/operations/list_loan_transactions.py
@@ -32,7 +32,7 @@ class ListLoanTransactionsResponse:
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""The request made is not valid."""
     loan_transactions: Optional[shared_loantransactions.LoanTransactions] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)