+++ previous-versions/assess/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.32.0] previous-versions/assess
 ### Releases
 - [PyPI v0.32.0] https://pypi.org/project/codat-assess/0.32.0 - previous-versions/assess
+
+## 2023-11-06 01:25:20
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.33.0] previous-versions/assess
+### Releases
+- [PyPI v0.33.0] https://pypi.org/project/codat-assess/0.33.0 - previous-versions/assess